+++ SleepHunter/Models/PlayerSortOrder.cs
@@ -1,0 +1,12 @@
+﻿
+namespace SleepHunter.Models
+{
+    public enum PlayerSortOrder
+    {
+        LoginTime,
+        Alphabetical,
+        HighestHealth,
+        HighestMana,
+        HighestCombined
+    }
+}